+++ packages/config/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## 2.2.14
+
+### Patch Changes
+
+- @latticexyz/common@2.2.14
+- @latticexyz/schema-type@2.2.14
+
 ## 2.2.13
 
 ### Patch Changes